up and running with TypeScript
